tests, clientcmd: Move CheckIfNoCmd to "checks"
The k8s reporter depends on clientcmd but this package has some extra
checks not realted to running commands and they depend on flag package
tha has a init function. This change move the check to the proper
framework/checks place.
