+++ tests/kubectl_test.go
@@ -31,7 +31,7 @@ var _ = Describe("[sig-compute]oc/kubectl integration", func() {
 	BeforeEach(func() {
 
 		k8sClient = clientcmd.GetK8sCmdClient()
-		clientcmd.SkipIfNoCmd(k8sClient)
+		checks.SkipIfNoCmd(k8sClient)
 	})
 
 	DescribeTable("[test_id:3812]explain vm/vmi", func(resource string) {
